add type conversion in rawposix

+++ src/example_grates/commonconstants.rs
@@ -2,7 +2,7 @@
 // instead of editing each implementation...
 
 /// Per-process maximum number of fds...
-pub const FD_PER_PROCESS_MAX: i32 = 1024;
+pub const FD_PER_PROCESS_MAX: u64 = 1024;
 
 /// Use this to indicate there isn't a real fd backing an item
 pub const NO_REAL_FD: u64 = 0xffabcdef01;
